Profiles added to schema, channel list shows which profile a channel is assigned to
In the original installation, each language production team had its own server and
had enough channels to warrant this. Now we have some languages with only a handful
of channels, yet they still require full production facilities, isolated from the
other production teams.

This update extends the Schema to support 'profiles', which will be used to
filter/restrict which channels are seen by each production team. Note this
restriction is not a security measure - it's just for practical purposes, so
that the different language production teams can have their own 'hotkey' buttons
and are not distracted by another team's GPI events.

This patch makes the channel list display which profile each channel is assigned to,
and uses CSS to make the channel list a little prettier.