Use sessions to pass around profile data
Was previously trying to avoid session data and make all requests
stateless, but it was getting silly having to pass around
profile_id or profile_code in urls, parameters or post data, and
also doing lots of profile name/code/id look-ups, so
I've given in and am converting the code to use sessions.

This patch sets the session variables in the root index method,
which then re-directs to the player. The player now displays the
current profile name.