assign_clips : comment out clip row delete link
This has been superceded by row select checkboxes and a 'delete selected'
link, which was requested in user feedback.

This makes it less easy to accidentally delete clips, but adds
an extra mouse click to any deletion. Only commented out to make it easy
to put back in if user opinion changes!