Add dummy argument to event type filter
For some reason certain combinations of event filters caused the
ajax http request to stop at readyState 2. Filter list ending in
'exported' was an example of this. This patch adds the argument
'dummy' to the end of the filter list to fix this problem. This
doesn't identify the root cause, but does solve the problem.