Update player.tt to work with newer versions of Chrome
Older versions of chrome only fire the player canplay event when we
load a new clip. Newer versions fire the canplay event every time we
jump to a new position in the audio file.

As we were doing initialisation of clip markers in the canplay event,
this caused previously marked in and out points to be discarded. It also
caused a problem with mouse interaction when locating an edit point.

The new behaviour is more standards-compliant, so should improve the experience
for users trying to use other browsers also.